Garden/requirements.txt
Django~=3.2.8
python-dotenv~=0.19.1
djangorestframework~=3.12.4